Created map lookup table from tilesheet
-Most deltaTime is now a double rather than a float
-Created a lookup table for the map
-Rendered a map to the screen according to the array of numbers

+++ RPG-Game/Bullet.cpp
@@ -10,7 +10,7 @@ Bullet::~Bullet()
 }
 
 
-void Bullet::Initialize(const sf::Vector2f &position, const sf::Vector2f &target, float speed)
+void Bullet::Initialize(const sf::Vector2f &position, const sf::Vector2f &target, double speed)
 {
 	this->speed = speed;
 	rectangleShape.setSize(sf::Vector2f(50, 25));
