+++ RPG-Game/FrameRate.cpp
@@ -21,7 +21,7 @@ void FrameRate::Load() {
 	}
 }
 
-void FrameRate::Update(float deltaTime) {
+void FrameRate::Update(double deltaTime) {
 	timer += deltaTime;
 
 	if (timer >= 100.0) {
